+++ src/main/java/webserver/RequestHandler.java
@@ -5,6 +5,8 @@ import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import utils.FileIoUtils;
+import webserver.request.HttpRequest;
+import webserver.request.HttpRequestParser;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
