+++ src/main/java/webserver/request/HttpRequest.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.request;
 
 import java.net.URI;
 import java.util.Optional;
